Character gets some damage on spike 'S'
