+++ sources/playerFile.h
@@ -14,5 +14,7 @@
 int addPlayerFile(Player *player, char *playerfile);
 void writePlayerFile(PlayerList playerList, char *playerfile);
 PlayerList *readPlayerFile(char *playerfile);
+Player * selectPlayer(char *playerfile);
+
 
 #endif // PLAYERFILE_H
